Add systemPropertyVariables to pom.xml Add ui, smoke, issue, link tags to checkCertificatePage.feature Add hooks package Move Hooks class to hooks package
