+++ src/test/java/cucumber/CucumberTestRunner.java
@@ -5,7 +5,8 @@ import io.cucumber.testng.CucumberOptions;
 
 @CucumberOptions(
         features = "src/test/java/cucumber/features",
-        plugin = {"pretty"}
+        plugin = {"pretty", "io.qameta.allure.cucumber7jvm.AllureCucumber7Jvm"},
+        tags = "@ui"
 )
 public class CucumberTestRunner extends AbstractTestNGCucumberTests {
 }
