+++ src/test/java/cucumber/features/checkCertificatePage.feature
@@ -1,5 +1,7 @@
+@ui
 Feature: Verify a certificate exists by number on CheckCertificate page
 
+  @smoke @issue=TEST-1 @link=https://example.org/issue/
   Scenario Outline:
     Given User wrote certificate number: <number> down
     Then  User clicked on the button and checks message by certificate number
