+++ src/test/java/cucumber/hooks/Hooks.java
@@ -1,10 +1,11 @@
-package cucumber.steps;
+package cucumber.hooks;
 
+import cucumber.steps.BaseUiSteps;
 import io.cucumber.java.AfterAll;
 
 public class Hooks extends BaseUiSteps {
 
-    @AfterAll
+    @AfterAll(order = Integer.MAX_VALUE)
     public static void afterAll() {
         driver.quit();
     }
