+++ src/test/java/cucumber/steps/CheckCertificatePageSteps.java
@@ -4,7 +4,6 @@ import com.hillel.pages.CheckCertificatePage;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import org.junit.jupiter.api.Assertions;
-import org.testng.annotations.Test;
 
 import static com.hillel.util.WaitUtils.waitUntilElementIsVisible;
 
@@ -12,14 +11,12 @@ public class CheckCertificatePageSteps extends BaseUiSteps {
     CheckCertificatePage checkCertificatePage = new CheckCertificatePage(driver);
 
     @Given("User wrote certificate number: {} down")
-    @Test(priority = 1)
     public void writeCertificateNumberDown(String number) {
         checkCertificatePage.open();
         checkCertificatePage.setNumberOfCertificate(number);
     }
 
     @Then("User clicked on the button and checks message by certificate number")
-    @Test(dependsOnMethods = {"writeCertificateNumberDown"})
     public void checkMessageIsDisplayed() {
         checkCertificatePage.getElCheckCertificateCta().click();
         waitUntilElementIsVisible(driver, checkCertificatePage.getElCertificateMessage());
