+++ src/test/resources/allure.properties
@@ -1,1 +1,2 @@
 allure.link.issue.pattern="https://example.org/issue/{}"
+allure.results.directory=target/allure-results
