+++ src/test/resources/testng.xml
@@ -1,5 +1,18 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+
+<!--<suite name="Practice Suite">
+    <packages>
+        <package name="testNg"></package>
+    </packages>
+    <test name="TestNgTest">
+        <groups>
+            <define name="smoke"></define>
+            <define name="regression"></define>
+        </groups>
+    </test> &lt;!&ndash; Test &ndash;&gt;
+</suite> &lt;!&ndash; Suite &ndash;&gt;-->
+
 <suite name="Practice Suite">
     <test name="TestNgTest">
         <classes>
